added minimal/infotext.lua
created a library to manage updating infotext.  it breaks lines up into
an array and builds an index based on keys using : as delimiter.

also moved loading of minimal/utility to top of list of dofiles so it
can be used in lower modules.  May want to move infotext.lua into
utilities later if its not too complicated.