fix for captured fish memeory and aqua_safe_path()
Fixed captured fish not remembering their age and energy level when
captured.

Added raycast() function to check full path is safe for fish.  Still
crashes into air, but think not as much. and gets stuck in corners -
think need to add a check for blocks above/below to see if change in y
would help find a path