Rewrote infotext to handle a fixed order list
rewrite and documented. Broke into sub functions to make it easier to
follow and reuse code snips.  Settled on keeping a list for unkeyed
lines to be added at the end of the keyed ones.  Keyed lines can
be replaced with new lines.

By default, the node description and owner if set are displayed as
line 1 and 2.  Order of other keyed lines can be set in a fixed order list.
Needed to insure multi-line infotext for things like transporters appear
in the correct order.  Calling set_infotext() without a list of lines to
set cause the node discription and owner to be set.