oversize-stacks-fix (#385)
This fixes the creation of oversize stacks of anperla from the bulk threshing recipe.

Crafting generally respects the stack size limits, and will start a new stack if the bulk recipe happens to go over the stack size ... but only if the bulk recipe itself doesn't produce more than the max stack size for that object.

Co-authored-by: Dan McNeill <danmcne@gmail.com>
Reviewed-on: https://codeberg.org/Mantar/Exile/pulls/385
Co-authored-by: ihaphleas <ihaphleas@noreply.codeberg.org>
Co-committed-by: ihaphleas <ihaphleas@noreply.codeberg.org>