Added "Lives" checker to prevent health effects from affecting players on respawn. Changed if statement on line 376 of mods/health/init.lua (#397)
Adds two new functions to mods/health/health_effects.lua

```lua
get_life_num(player)
```
returns the player's meta int "lives" - used as a ID system of sorts (suggested by someone once I think, or maybe I thought of it? I don't remember)

```lua
is_illness_valid(player,life_num)
```
returns true if the select health_effect that is supposed to occur in a minetest.after() is allowed to run (current only condition for being true is if the player's "lives" value is still equal to the "lives" value that was registered upon the setting of the health effect)
otherwise returns false (which prevents the health effect from occurring as I program it to "return" if it's false)

I intend to use this function more extensively with more options when I get to reworking health_effects, and thus give it this general name. For example, I want this function to be able to prevent health effects affecting the player if enable_damage is set to false instead of having the health/init.lua code constantly set the player's stats to full

functions are lines 55 to 88 of the file

I made some small modifications to line 376 of mods/health/init.lua
```lua
if temperature > 100 or temperature < 0 then
```
to
```lua
if temperature >= 100 or temperature <= 0 then -- now will cause immediate death
```
not sure if it'd cause immediate death, but if the body temperature is greater than or equal to 100, or lesser than or equal to 0, it will cause whatever effects are listed there, to occur

Co-authored-by: TPH <DamotrixRob@gmail.com>
Reviewed-on: https://codeberg.org/Mantar/Exile/pulls/397
Co-authored-by: TPH <tph@noreply.codeberg.org>
Co-committed-by: TPH <tph@noreply.codeberg.org>