Hacky Fixes for Watering Can Fixes #424 (#431)
This basically checks if the provided sediment's name has a "depleted" or "roots", and renames the node in accordance to how they are named in-game to get the wet version.

Also a fix for #430 provided if the code inside, on lines 713-717 (checks only wet nodes under selected plants) and lines 720-724 (will check if selected node is wet and will not dump water on the node) have their notation removed. Remove the entirety of lines 713-717 (plant-only) IF lines 720-724 (sediment) are used.

Reviewed-on: https://codeberg.org/Mantar/Exile/pulls/431
Co-authored-by: TPH <DamotrixRob@gmail.com>
Co-committed-by: TPH <DamotrixRob@gmail.com>