Small Animal Probe Fix (#440)
I've gotten an error for "ent" being nil atleast twice I think (I confirmed it once when I was burning a Darkasthaan in lava and kept constantly clicking on it with animal_probe). And I decided to use minetest's built-in "is_player()" function instead of depending on pointed_thing's entity ref in case of a special case of where is_player() is not a valid function of what is pointed at.

Reviewed-on: https://codeberg.org/Mantar/Exile/pulls/440
Co-authored-by: TPH <tph@noreply.codeberg.org>
Co-committed-by: TPH <tph@noreply.codeberg.org>