HUD overlay: Add visual aid: quickslot numbers
To overcome the constant frustration of having to count out quickslots
visually to press the right number key or wait & double check to see if
the scroll-wheel encoder of the mouse did not misread the scroll (again)
-- a physical/optic glitch -- and jumped to a further slot or, after
stopping the scroll, jumped a slot back, to ensure not performing
actions with the wrong object/block, which might have grave consequences
to a game that was invested a lot of time & effort in.
  This constant checking and double-checking is considered needlessly
spent time & mental effort, and taking away from game immersion.

The understanding of the commit author suggests that the game was NOT
meant to be challenging in the way of users interacting with game
controls.
  Even if it is supposed to be part of it, there is still ample
opportunity for the user to mis-press the key if they are too quick
about it, however, this time it was clear what they should have and
meant to press, and it is reasonable to think that the frustration will
not be directed towards the game, but their own carelessness.