mods/minimal/init.lua: Fix indentation whitespace to be uniform
To fix line-start misalignment for people not exactly having the same
tab-length settings that would make space-indent and tab-indent on
various lines line up.
Making reading the code easier for everyone.

Was predominantly tabs, so made it all tabs.

Additionally, if there is a code-style rule to use space-indent only, now it should be
the matter of 1 command to fix it.

Also, removed empty line at EOF.