Fix: we were not executing first_spawn() (#697)
6d960f339e6961aecc4acf91ead563b6b70d49ee Introduced a bug:
1. New players were invisible
2. Portal effects did not trigger when new players spawned
3. Energy, thirst and other attributes did not reset after reading the lore popup

I was able to reproduce the bug:
1. In single-player - always
2. In multi-player - if the message of the day is empty

The cause is that the queue assumed show_motd always shows a formspec, but show_motd sometimes returned early.
The fix: the functions we put in the queue now tell the queue whether they are done or not

I also cleaned up some code that looks like leftovers from debugging.

Reviewed-on: https://codeberg.org/Mantar/Exile/pulls/697
Co-authored-by: hexagon-recursion <hexagon-recursion@posteo.net>
Co-committed-by: hexagon-recursion <hexagon-recursion@posteo.net>