Updated tooltips to save as def._tool_tips instead of modifying description
Desription set via metadata in crafting/api.lua when items are crafted instead.
Does away with then to track old description; original description can be pulled from
the definintion and modified description on the item from the item meta data.
mouseover is updated using metedata if found.

Was needed to add creator to tools and keys in their description.