Disappearing Fruit Resolved Fixes #563 (#767)
Seems it also produced a tree mark when the fruit was placed? Meaning that it would infinitely regrow the fruit, maybe? As long as there was wood or leaves I guess LOL

after_place_node ran set_node which caused the after_destruct to be ran twice - maybe something to do with how the tree marks work?

My resolution was just having it use switch_node instead, no issues there.

Wasn't sure if creative players should be able to place wild tree fruits so I left it as is.

Reviewed-on: https://codeberg.org/Mantar/Exile/pulls/767
Co-authored-by: TPH <DamotrixRob@gmail.com>
Co-committed-by: TPH <DamotrixRob@gmail.com>