Crafting - 'Use only': hints on by default / button available through setting
We had a hint button with 'Use only' input mode to highlight possible recipes.
When opening the formspec hints where off by default because of the incomplete
background refreches. Without cradfting in inventory proper highlighting
can always be provided. The old behaviour with a button remains available for players
who want this part to be more difficult.