crafting: With 'stack' selected, crafting fails for some recipes although there is enough inputs.
What was wrong?
The calculation could return fractional numbers of outputs to be crafted, which may in turn require a fractional number of inputs.
Example: For crafting a stack of `dry stack`, which would be 3, the function calculate_stack_input()
returned 1.5 which was not only used in tooltips but also to actually attempt to craft a recipe 1.5 times,
resulting in several lines of warnings in the chat.
Now the funktion applies math.floor() before returning its prior result.