to check: paint (graffiti) behavior improvement (linked to issue #332 from 2 years ago) (#1420)
So, the goal here is to have up/down/left/right arrow correctly doing that on the floor, and not north/south/west/east.

Which is achieved here by changing the node from wallmounted type (doesn't let us do horizontal rotation on the floor, if I understood it well, not enough possible rotations) to facedir type (more rotations).

Reviewed-on: https://codeberg.org/Mantar/Exile/pulls/1420
Co-authored-by: lili36 <lili36@noreply.codeberg.org>
Co-committed-by: lili36 <lili36@noreply.codeberg.org>